Re-order package.json exports for Webpack 5 browser usage
The export conditions are evaluated and the first that
matches will be used. The ./nodejs/filecoin_signer_wasm.js
gets matched for "import" or "require", so the the "browser"
condition with ./browser/filecoin_signer_wasm.js never gets
matched. Webpack 4 automatically installs polyfills for Node.js
libraries but Webpack 5 won't.

https://webpack.js.org/guides/package-exports/

Re-ordering the export conditions so that the browser condition
appears first allows this library to be used with Webpack 5
for browser builds without polyfills.